+++ dblib/tns_dns_docker_test.go
@@ -1,8 +1,10 @@
 package dblib
 
 import (
+	"context"
 	"fmt"
 	"net"
+	"net/netip"
 	"os"
 	"time"
 
@@ -10,8 +12,9 @@ import (
 	"github.com/tommi2day/gomodules/netlib"
 	"github.com/tommi2day/gomodules/test"
 
-	"github.com/ory/dockertest/v3"
-	"github.com/ory/dockertest/v3/docker"
+	"github.com/moby/moby/api/types/container"
+	"github.com/moby/moby/api/types/network"
+	"github.com/ory/dockertest/v4"
 )
 
 const (
@@ -25,21 +28,20 @@ const (
 
 var (
 	dblibDNSContainerName  string
-	dblibDNSContainer      *dockertest.Resource
-	dblibDNSNetwork        *dockertest.Network
+	dblibDNSContainer      dockertest.ClosableResource
+	dblibDNSNetwork        dockertest.ClosableNetwork
 	dblibDNSNetworkCreated = false
 	dblibDNSServer         = "127.0.0.1"
 )
 
 // prepareDBlibDNSContainer create a Bind9 Docker Container
-func prepareDBlibDNSContainer() (container *dockertest.Resource, err error) {
+func prepareDBlibDNSContainer() (resource dockertest.ClosableResource, err error) {
 	if os.Getenv("SKIP_DB_DNS") != "" {
 		return nil, fmt.Errorf("skipping DB DNS Container in CI environment")
 	}
 
 	dblibDNSContainerName = getContainerName()
-	// use versioned docker pool because of api error client version to old
-	pool, err := common.GetVersionedDockerPool("")
+	pool, err := common.GetDockerPool()
 	if err != nil {
 		return nil, err
 	}
@@ -49,14 +51,14 @@ func prepareDBlibDNSContainer() (container *dockertest.Resource, err error) {
 		return nil, err
 	}
 
-	container, err = buildAndRunContainer(pool)
+	resource, err = buildAndRunContainer(pool)
 	if err != nil {
 		return
 	}
 
 	time.Sleep(10 * time.Second)
 
-	ip := validateContainerIP(container)
+	ip := validateContainerIP(resource)
 	if ip == "" {
 		return
 	}
@@ -78,29 +80,9 @@ func getContainerName() string {
 	return name
 }
 
-func setupNetwork(pool *dockertest.Pool) error {
-	networks, err := pool.NetworksByName(dblibNetworkName)
-	if err != nil || len(networks) == 0 {
-		return createNetwork(pool)
-	}
-	dblibDNSNetwork = &networks[0]
-	return nil
-}
-
-func createNetwork(pool *dockertest.Pool) error {
+func setupNetwork(pool dockertest.Pool) error {
 	var err error
-	dblibDNSNetwork, err = pool.CreateNetwork(dblibNetworkName, func(options *docker.CreateNetworkOptions) {
-		options.Name = dblibNetworkName
-		options.CheckDuplicate = true
-		options.IPAM = &docker.IPAMOptions{
-			Driver: "default",
-			Config: []docker.IPAMConfig{{
-				Subnet:  dblibNetworkPrefix + ".0/24",
-				Gateway: dblibNetworkPrefix + ".1",
-			}},
-		}
-		options.EnableIPv6 = false
-	})
+	dblibDNSNetwork, err = common.CreateNetworkWithSubnet(pool, dblibNetworkName, dblibNetworkPrefix+".0/24", dblibNetworkPrefix+".1")
 	if err != nil {
 		return fmt.Errorf("could not create Network: %s:%s", dblibNetworkName, err)
 	}
@@ -108,46 +90,60 @@ func createNetwork(pool *dockertest.Pool) error {
 	return nil
 }
 
-func buildAndRunContainer(pool *dockertest.Pool) (*dockertest.Resource, error) {
+func buildAndRunContainer(pool dockertest.Pool) (dockertest.ClosableResource, error) {
 	vendorImagePrefix := os.Getenv("VENDOR_IMAGE_PREFIX")
 	fmt.Printf("Try to build and start docker container %s\n", dblibDNSContainerName)
-	buildArgs := []docker.BuildArg{
-		{Name: "VENDOR_IMAGE_PREFIX", Value: vendorImagePrefix},
-		{Name: "BIND9_VERSION", Value: dblibRepoTag},
+	repoTagArg := dblibRepoTag
+	buildArgs := map[string]*string{
+		"VENDOR_IMAGE_PREFIX": &vendorImagePrefix,
+		"BIND9_VERSION":       &repoTagArg,
 	}
 
+	ctx := context.Background()
 	dockerContextDir := test.TestDir + "/docker/oracle-dns"
-	return pool.BuildAndRunWithBuildOptions(
+	resource, err := pool.BuildAndRun(ctx, "dblib-bind9:test",
 		&dockertest.BuildOptions{
 			BuildArgs:  buildArgs,
 			ContextDir: dockerContextDir,
 			Dockerfile: "Dockerfile",
 		},
-		&dockertest.RunOptions{
-			Hostname:     dblibDNSContainerName,
-			Name:         dblibDNSContainerName,
-			Networks:     []*dockertest.Network{dblibDNSNetwork},
-			ExposedPorts: []string{"9054/tcp"},
-			// need fixed mapping here
-			PortBindings: map[docker.Port][]docker.PortBinding{
-				"9054/tcp": {
-					{HostIP: "0.0.0.0", HostPort: fmt.Sprintf("%d", dblibDNSPort)},
-				},
-			},
-		}, func(config *docker.HostConfig) {
+		dockertest.WithHostname(dblibDNSContainerName),
+		dockertest.WithName(dblibDNSContainerName),
+		dockertest.WithContainerConfig(func(config *container.Config) {
+			if config.ExposedPorts == nil {
+				config.ExposedPorts = network.PortSet{}
+			}
+			config.ExposedPorts[network.MustParsePort("9054/tcp")] = struct{}{}
+		}),
+		dockertest.WithHostConfig(func(config *container.HostConfig) {
 			config.AutoRemove = false
-			config.RestartPolicy = docker.RestartPolicy{Name: "no"}
-		})
+			config.RestartPolicy = container.RestartPolicy{Name: container.RestartPolicyDisabled}
+			// need fixed mapping here
+			config.PortBindings = network.PortMap{
+				network.MustParsePort("9054/tcp"): {
+					{HostIP: netip.MustParseAddr("0.0.0.0"), HostPort: fmt.Sprintf("%d", dblibDNSPort)},
+				},
+			}
+		}),
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	if err = resource.ConnectToNetwork(ctx, dblibDNSNetwork); err != nil {
+		return nil, fmt.Errorf("could not connect container to network %s: %w", dblibNetworkName, err)
+	}
+	return resource, nil
 }
 
-func validateContainerIP(container *dockertest.Resource) string {
-	ip := container.GetIPInNetwork(dblibDNSNetwork)
+func validateContainerIP(resource dockertest.Resource) string {
+	ip := resource.GetIPInNetwork(dblibDNSNetwork)
 	fmt.Printf("DB DNS Container IP: %s\n", ip)
 	return ip
 }
 
-func waitForDNSServer(pool *dockertest.Pool) error {
-	dh := common.GetDockerHost(pool)
+func waitForDNSServer(pool dockertest.Pool) error {
+	dh := common.GetDockerHost(pool.Client().DaemonHost())
 	if dh != "" {
 		fmt.Printf("Docker Host: %s\n", dh)
 	}
@@ -171,9 +167,8 @@ func waitForDNSServer(pool *dockertest.Pool) error {
 	ip := lips[0]
 	dblibDNSServer = ns
 	fmt.Printf("DNS Host %s  IP resolved as %s\n", dblibDNSServer, ip)
-	pool.MaxWait = dblibDNSContainerTimeout * time.Second
 	start := time.Now()
-	err = pool.Retry(func() error {
+	err = pool.Retry(context.Background(), dblibDNSContainerTimeout*time.Second, func() error {
 		c, e := net.Dial("tcp", net.JoinHostPort(dblibDNSServer, fmt.Sprintf("%d", dblibDNSPort)))
 		if e != nil {
 			fmt.Printf("Err:%s\n", e)
@@ -207,12 +202,13 @@ func testDNSResolution() error {
 	return nil
 }
 
-func destroyDNSContainer(container *dockertest.Resource) {
-	if container != nil {
-		_ = container.Close()
+func destroyDNSContainer(resource dockertest.ClosableResource) {
+	ctx := context.Background()
+	if resource != nil {
+		_ = resource.Close(ctx)
 	}
 
 	if dblibDNSNetworkCreated && dblibDNSNetwork != nil {
-		_ = dblibDNSNetwork.Close()
+		_ = dblibDNSNetwork.Close(ctx)
 	}
 }