+++ pwlib/vault_docker_test.go
@@ -1,6 +1,7 @@
 package pwlib
 
 import (
+	"context"
 	"fmt"
 	"net/http"
 	"os"
@@ -10,8 +11,8 @@ import (
 
 	"github.com/tommi2day/gomodules/test"
 
-	"github.com/ory/dockertest/v3"
-	"github.com/ory/dockertest/v3/docker"
+	"github.com/moby/moby/api/types/container"
+	"github.com/ory/dockertest/v4"
 )
 
 const repo = "docker.io/hashicorp/vault"
@@ -22,7 +23,7 @@ const rootToken = "pwlib-test"
 var containerName string
 
 // prepareVaultContainer create an Oracle Docker Container
-func prepareVaultContainer() (container *dockertest.Resource, err error) {
+func prepareVaultContainer() (resource dockertest.ClosableResource, err error) {
 	if os.Getenv("SKIP_VAULT") != "" {
 		err = fmt.Errorf("skipping Vault Container in CI environment")
 		return
@@ -40,38 +41,38 @@ func prepareVaultContainer() (container *dockertest.Resource, err error) {
 	vendorImagePrefix := os.Getenv("VENDOR_IMAGE_PREFIX")
 	repoString := vendorImagePrefix + repo
 
+	ctx := context.Background()
 	fmt.Printf("Try to start docker container for %s:%s\n", repoString, repoTag)
-	container, err = pool.RunWithOptions(&dockertest.RunOptions{
-		Repository: repoString,
-		Tag:        repoTag,
-		Env: []string{
+	resource, err = pool.Run(ctx, repoString,
+		dockertest.WithTag(repoTag),
+		dockertest.WithEnv([]string{
 			"VAULT_DEV_ROOT_TOKEN_ID=" + rootToken,
 			"VAULT_DEV_LISTEN_ADDRESS=0.0.0.0:8200",
-		},
-		Hostname: containerName,
-		Name:     containerName,
-		CapAdd:   []string{"IPC_LOCK"},
-		Cmd:      []string{},
-		Mounts: []string{
+		}),
+		dockertest.WithHostname(containerName),
+		dockertest.WithName(containerName),
+		dockertest.WithCmd([]string{}),
+		dockertest.WithMounts([]string{
 			test.TestDir + "/docker/vault_provision:/vault_provision/",
-		},
-	}, func(config *docker.HostConfig) {
-		// set AutoRemove to true so that stopped container goes away by itself
-		config.AutoRemove = true
-		config.RestartPolicy = docker.RestartPolicy{Name: "no"}
-	})
+		}),
+		dockertest.WithHostConfig(func(config *container.HostConfig) {
+			// set AutoRemove to true so that stopped container goes away by itself
+			config.AutoRemove = true
+			config.RestartPolicy = container.RestartPolicy{Name: container.RestartPolicyDisabled}
+			config.CapAdd = []string{"IPC_LOCK"}
+		}),
+	)
 
-	if err != nil || container == nil {
+	if err != nil || resource == nil {
 		err = fmt.Errorf("error starting vault docker container: %v", err)
 		return
 	}
 
-	pool.MaxWait = containerTimeout * time.Second
-	host, port := common.GetContainerHostAndPort(container, "8200/tcp")
+	host, port := common.GetContainerHostAndPort(resource, "8200/tcp")
 	address := fmt.Sprintf("http://%s:%d", host, port)
 	fmt.Printf("Wait to successfully connect to Vault with %s (max %ds)...\n", address, containerTimeout)
 	start := time.Now()
-	if err = pool.Retry(func() error {
+	if err = pool.Retry(ctx, containerTimeout*time.Second, func() error {
 		var resp *http.Response
 		//nolint gosec
 		resp, err = http.Get(address)
@@ -95,7 +96,7 @@ func prepareVaultContainer() (container *dockertest.Resource, err error) {
 	// provision
 	cmdout := ""
 	cmd := []string{"bash /vault_provision/vault_init.sh"}
-	cmdout, _, err = common.ExecDockerCmd(container, cmd)
+	cmdout, _, err = common.ExecDockerCmd(resource, cmd)
 	if err != nil {
 		fmt.Printf("Exec Error %s", err)
 	} else {